requirements-dev.txt
attrs >=17.4
black ==24.3.0
bump-my-version
compas_invocations2
compas_viewer
invoke >=0.14
ruff
sphinx_compas2_theme
twine
wheel
line_profiler
scikit-build-core[pyproject] >=0.10
twine
wheel
setuptools
nanobind
cibuildwheel==2.23.1